fix: guard delete_fn shutdown panic, add wait_for_all_merges, u64 merge sum
- CORR-6: Guard SegmentSnapshot delete_fn with Handle::try_current()
  to prevent panic when tokio runtime is gone during program shutdown
- CORR-7: Split wait_for_merging_thread (single wait) from new
  wait_for_all_merges (loops with merge policy re-evaluation)
- CORR-8: Use u64 for total_docs sum in TieredMergePolicy::find_merge
  to prevent theoretical u32 overflow

+++ hermes-core/src/merge/mod.rs
@@ -149,8 +149,8 @@ impl MergePolicy for TieredMergePolicy {
 
                 let chunk = &sorted[..sorted.len().min(self.max_merge_at_once)];
                 if chunk.len() >= 2 {
-                    let total_docs: u32 = chunk.iter().map(|s| s.num_docs).sum();
-                    if total_docs <= self.max_merged_docs {
+                    let total_docs: u64 = chunk.iter().map(|s| s.num_docs as u64).sum();
+                    if total_docs <= self.max_merged_docs as u64 {
                         return Some(MergeCandidate {
                             segment_ids: chunk.iter().map(|s| s.id.clone()).collect(),
                         });
